update put delete

+++ src/item/item.service.ts
@@ -15,6 +15,13 @@ export class ItemService {
     return await this.itemRepository.find();
   }
 
+  async findPublicAll(): Promise<Item[]> {
+    const items = await this.itemRepository.find();
+    return items.map((x) => {
+      return { id: x.id, todo: x.todo, limit: x.limit, isDone: x.isDone };
+    });
+  }
+
   async create(item: CreateItemDTO): Promise<InsertResult> {
     return await this.itemRepository.insert(item);
   }
@@ -23,11 +30,25 @@ export class ItemService {
     return await this.itemRepository.findOne({ id: id });
   }
 
-  async update(id: number, item: Item): Promise<UpdateResult> {
+  async update(id: number, item): Promise<UpdateResult> {
     return await this.itemRepository.update(id, item);
   }
 
   async delete(id: number): Promise<DeleteResult> {
+    return await this.itemRepository.delete(id);
+  }
+
+  async deleteByPassword(
+    id: number,
+    deletePassword: string,
+  ): Promise<DeleteResult> {
+    const targetItem = await this.find(id);
+    if (!targetItem) {
+      return Promise.reject(new Error("Missing Item."));
+    }
+    if (targetItem.deletePassword !== deletePassword) {
+      return Promise.reject(new Error("Incorrect password"));
+    }
     return await this.itemRepository.delete(id);
   }
 }
